fuzz: remove 'honggfuzz' feature, clean up warnings
We only ever fuzz with honggfuzz, so make this unconditional. Meanwhile
update the fuzz harness to 2018 so we can drop the 'extern crate's, fix
the honggfuzz version at 0.5.55 (earlier ones do not work), and fix
warnings.

This simplifies CI a bit; a later PR will bring the new fuzz scripts in
from rust-bitcoin.