inference: refactor/rewrite `for_node` (was `get_arrow`)
The existing type inference code had a lot of calls to bind and unify
that were marked "cannot fail" because we were binding/unifying free
variables with no constrains.

When binding to a free variable we should just replace the free variable
with Variable::bound(); when unifying we should use a new constructor to
construct a VariableInner::EqualTo, or we can "cheat" and just create a
new refcounted pointer to the original variable.

After doing these transformations it was clear that most combinators
actually cannot fail type unification; it's only the four binary ones
that can (comp, pair, case, disconnect). So refactor out the
constructors for the non-fallible ones so that users can call them
without unwrapping, and then refactor `for_node` to call these.

Finally, rearrange the logic for the binary ones to just match on the
children rather than calling get_left/get_right, and for `Case` make the
case/assertl/assertr logic more direct. These slightly lengthen the code
but reduces the amount of indentation and make it much clearer what's
going on.