types: introduce new `Precomputed` variant of `VariableInner`
In the following commits we are going to try to store precomputed
variables in the context object. This will cause problems in the case
that the precomputed variables get finalized, since during finalization,
we lose information about the original RcVar.

Trying to prevent finalization of the precomputed types, aside from not
clearly being possible, would be inefficient since the result would be
repeated finalization of the same precomputed types.

Instead we treat precomputed types as always being both finalized and
unfinalized, and modify the union-bound algorithm to make sure that the
precomp types are never modified during unification. This change is a
little bit subtle and we include a long comment which (hopefully)
justifies it.