ffi: add `ffi` module with updated correct bindings to C functions
Adds new type aliases `UWORD` and `ubounded` and check that their size
and alignment match those of the corresponding C structures. Moves the
size/alignment constants to the new ffi.rs file and patches up a couple
other references.

The next commit will introduce similar size/alignment tests for all of
the other structures.

We'll try to add enough tests etc that we'll notice when the function
bindings go out of sync, and by having a single file and a consistent
naming scheme, we should be okay.

Arguably we should auto-generate these in our re-vendoring script.