dag: replace iteration algorithm with one that tracks indices
The current iteration algorithm uses the `visited` set to avoid
repeating nodes; when we change this to use a more abstract notion
of equality, this will stop working (resulting in infinite loops
whenever we want to yield nodes twice, as we will do when unsharing
nodes).

This algorithm is also unable to refer to yielded items' children by
their indices, which will be necessary when we implement an unshared
iterator, since otherwise users of the iterator will have a very
difficult time distinguishing between different times that the same
node is yielded.

This results in a nontrivial increase in complexity but unfortunately
I think it's a necessary one.