commit: refactor `CommitNode::finalize` to use new iteration abstraction
`CommitNode::finalize` has some unique requirements relative to other
iterators. In particular, it is an iterator over `CommitNode`s which
nonetheless attempts to share based on IMRs, which it is computing in
parallel to the actual iteration.

Furthermore, when attaching fresh witnesses, we want to unshare the
witnesses (and all their parents) even if they had originally been
shared in the `CommitNode` DAG.

The old logic did this in an ad-hoc way and didn't actually work -- if a
witness was part of a larger tree that was shared, we wouldn't detect
this and unshare the tree.

Now we instead allow iteration with a new "unshare witnesses and disconnect"
adaptor, which hides all the messy logic of unsharing witnesses and their
parents within dag.rs. We apply that adaptor to a new special-purpose
iteration tracker within `finalize()`, which computes IMRs and finalized
nodes as the iteration progresses.