dag: use sharing iterators to replace `sharing` module
This lets us replace the `sharing` module; to share, we just iterate
over the program with the "sharing" iterator, and to check sharing, we
check the "internal sharing" iterator (which reflects the actual Rc
structure) against the "sharing" one.

It leaves `CommitNode::finalize` alone since dealing with this use of
iterators will be a little bit more complicated. It will be addressed
in the next three commits.