decode: replace decoding algorithm with a 2-stage algorithm that doesn't use CommitNode::Hidden
The new algorithm is longer but IMHO a bit more elegant. We first decode
the bitstream into a node structure that contains only indices; then we
map this into a CommitNode structure, checking the conditions on hidden
nodes as we go.

Thanks to our existing PostOrderIter abstraction the canonical-ordering
check continues to be more-or-less trivial.