decode: move decode_value into method on BitIter
While we are at it, replace the recursive implementation with an
iterative one.

This looks like it's doing two distinct things, introducing this new
`BitIter::byte_slice_window` and also porting `decode_value`, but
actually they're related since I need `byte_slice_window` to replace the
use of `decode_value` in the executor. I could split these commits up
but it'd be a bit ugly/hacky.