analysis: implement a bunch of constructors for NodeBounds
Should be easy to verify that all the new code matches the existing
code. (Whether it's all correct is a problem for a later PR. :))

While I'm at it, move `NodeBounds` to the `analysis` module and add a
bunch more #derives so that it can be used in various collections. It's
a POD type so I'm not too worried about people being confused about the
meaning of these.