dag: introduce right-to-left post order iterator
There is a very common pattern in Simplicity where we have a program
fragment of the form comp(witness, case(...)). When iterating, we will
yield first the witness, then the case's children, then the case itself.

But because the witness decides which child is going to be hidden, it
would be convenient if the children were yielded first. We can achieve
this with a small amount of potent magic: we can just relabel the
children and then use the normal post-order iterator, and relabel the
children that are yielded.

In practice this is all encapsulated within the new `rtl_post_order_iter`
method.

This is probably a general pattern when satisfying Simplicity programs,
that we would like to consider the right branch of a 'comp' before
deciding what to do with the left branch. Ditto for disconnect.