decode: remove public export of `decode_program`
In the next commits we will export the new `CommitNode::decode` which
will replace `decode::decode_program` (decoding a witness-less program
while doing typechecking and checking that the expression is 1->1) as
well as `RedeemNode::decode`.

Internally, we will continue to have `decode_expression` which will be
useful for unit tests, as it will allow decoding arbitrary expressions
from a bit encoding (with typechecking, but without sharing checks or
a 1->1 check), but we will not expose such a thing.