dag: fix `PostOrderIter::into_display`
This hasn't been updated since we first introduced the `dag` module, and
it basically has legacy code that was translated well enough to compile;
it assumes `InternalSharing` and manually tracks indices. The first can
cause crashing bugs and the latter is unnecessary.

Rewrite this function to be tighter, work for arbitrary sharing choices,
and to no longer have any `unwrap`s.