satisfier: static lookup instead of dynamic signining
The previous satisfier dynamically signed the transaction when queried.
This required it to store an Elements environment.
Not only is this ugly, but the hardcoded control block is actually invalid for arbitrary internal keys!
The new design turns the satisfier into a static lookup table.
We can convert it into a PSET plus index later if needed.