node: split `Constructible` up into three traits
Currently the `W` parameter is only used by `witness()` which causes
type ambiguity in some cases; later `J` will be only used by `jet()`
which will be even worse. (Currently this is not a problem for W because
we're careful to always implement `Constructible` for a single specific
value of W, but that's annoying to keep doing, and it's not a problem
for J because we take a `Context<J>` as a parameter to every method,
so J is forced by that. But in the next commit we will drop the `<J>`
from `Context` and we'll start to get problems everywhere.)