context: drop <J> parameter from Context
We had originally imagined that the context object would serve as a
global cache of Simplicity fragments so that whenever we constructed
something, we might simply look up each fragment in the cache, and
reuse that to save memory.

However, this is something of a pipe dream for several reasons:

* When constructing ConstructNodes, the user can just do this himself
  because every step of the construction is explicit. But he may not
  want to, because type inference can cause weird effects where types
  from one program affect types in another.

* When constructing CommitNodes or RedeemNodes directly, the user will
  be starting with some sort of encoded version of the program, which
  will be heavily shared, so we can save space by respecting the encoded
  sharing, and the marginal benefit from global sharing is pretty small.

* In all cases it's unclear whether the cost of a hashmap lookup or the
  reduced locality are a reasonable tradeoff for the memory savings.

* If we really wanted to do this, we could implement it by creating a
  thread-local global cache or something and then it would not affect
  the API for normal users.

For these reasons and more, we're probably never going to put a node
cache in the Context object. So we don't need the <J> parameter, so it's
just polluting the API. Let's get rid of it.