node: add `SimpleFinalizer` type for testing and basic cases
This also demonstrates how to use the `Finalizer` trait. It takes a
witness iterator, and provides witnesses when requested; if it runs out
of witnesses it stops providing them, and will then start pruning nodes.

This commit also deletes the `finalize_no_witnesses` method on
`CommitNode` which is now pretty-much redundant. It's easy to simulate
and only useful for testing.