types: use global atomic for names for anonymous types
Rather than having a context object that must be threaded through the
entire library, it is easier to just use an atomic. Users who want
precise control over type names are already lacking this, because they
can't tell in what order names are produced, and users who don't want
the control are burdened by the context object.

If users *do* want to rename types, they can currently do this in an
ugly way by calling Type::free and then binding the result to the type
they want, which will work if the target type is free. In future, we
can provide a saner API, such as an alternate set of constructors for
ConstructNode. But the current API is more burden than it's worth.

This reduces the `Context` object's role to only providing precomputed
power-of-two types.