decode: switch `decode` module to use the new types
Now `decode_expression` produces a `ConstructNode` while
`decode_program` produces a `CommitNode`.

While we are transitioning to the new API, we introduce some scaffolding
to convert `ConstructNode`s directly to the old-style `CommitNode`s.
Since `ConstructNode` does not do a 1->1 check, we (temporarily) lose
that check for both `CommitNode` and `RedeemNode`. Unit tests do not
fail (we should revisit this..) but the C-vs-Rust fuzz tests will.

This is okay, for now. For the new-style nodes, this check is *always*
done when converting from a `ConstructNode` to a `CommitNode`, and the
old-style nodes are going to be deleted. Right now no unit tests fail
because of this difference; we should add some later that make sure
that the 1->1 check is done for both CommitNodes and RedeemNodes.

Right now sharing is enforced for `RedeemNode`s but not `CommitNode`s.
The next commits will enforce a form of sharing for `CommitNode`: that
we have maximal sharing except that witness and disconnect nodes are
never equal to each other, so they (and their parents) are never shared.