policy: encapsulate the enum in a new `Fragment` type
The existing `Policy` type is a bare enum of the different policy
fragments. This prevents us from attaching any metadata. Rename the
existing `Policy` type to `Fragment`, wrap it in a new `Policy` struct,
and create forwarding methods for all the constructors and the
translation function.

This commit simply drops `Policy::sort` and `Policy::normalize`. These
methods get very limited (no?) use in rust-miniscript and make even less
sense for Simplicity than they do for Miniscript. Their purpose is to do
comparisons between policies that are identical up to ordering and some
simple transformations. If we decide that we want this functionality, we
can implement it in a more direct way.