policy: make `and` and `or` binary operations
We have currently typed And ond Or as taking vectors of children, but we
then assert that there are always exactly two. We should just use the
type system to enforce this, and save some vector allocations and
assertions.

Also deletes two FIXME comments in policy/embed.rs. These comments refer
to rearranging Miniscript integration code, in poorly-specified and
unnecessary ways. Maybe we will revisit this module in the future to
clean it up but there is nothing to "fix" and having FIXME comments
distracts from actual FIXMEs.