policy: require ToPublicKey on all keys; drop PublicKey32
This loses us the ability to have abstract key types in policy fragments
(e.g. strings) but unfortunately I don't see a way to keep that while
maintaining the "policies are 1<->1 with Simplicity fragments" model. At
least, policies with generic keys cannot be converted to Simplicity and
cannot be satisfied so it's not clear what they'd even be for.

Eventually we may want to introduce an actual abstract policy type,
which can be compiled into our existing "policy" if its keytype
satisfied ToPublicKey (or maybe a weaker PublicKey32-like trait).

Also, drop our PublicKey32 trait. As written, it is only possible to
implement for XOnlyPublicKey32 because no other key type in use can
round-trip from 32-bytes. And it's only actually used in one place: to
load an unspendable key for use with Taproot. So replace that single use
with a dedicated trait.

By dropping this we gain the ability to use xpubs everywhere, at least
if we were to implement the Unspendable trait on them (which we ought
to).