node: introduce `CommitNode::finalize` and `SimpleFinalizer`
Adds the `CommitNode::finalize` method which basically just wraps
`Node::convert` and fixes the sharing choice to `NoSharing`.

Also adds a new `SimpleFinalizer` struct, which demonstrates how to use
the `Converter` trait to do finalization. It takes a witness iterator
and provides witnesses when requested; if it runs out of witnesses it
returns an error.

This commit also deletes the `finalize_no_witnesses` method on
`CommitNode` which is now pretty-much redundant.