encode: implement `encode` methods directly on *Node types
This also contains a significant functionality change: when encoding
CommitNodes in the bit encoding, witnesses and disconnect nodes will be
unshared. This is consistent with the previous commit, but previously,
when encoding `CommitNode`s we would just use whatever internal sharing
we happened to have.

This, and the fact that we no longer define an encoding for non-1->1
`ConstructNode`s, mean some unit tests need to be adjusted.