node: introduce WitnessNode which is used to construct programs with witnesses
This node type supports having witnesses or making them as "pruned"
meaning that no witness is available and the node should be pruned. It
then has a method `prune_and_retype` which converts case nodes to
asserts to actually prune any branches that need to be.

Pruning may remove type constraints and reduce the size of types,
including those of witness types; in this case we need to shrink the
witnesses. This PR does not introduce that support, it just leaves a
FIXME about it.

It also has a FIXME to do a final bit-machine-based pruning pass before
finalizing the program.