jet: remove a bunch of dead code
We have some ancient pure Rust implementations of a small set of jets.
These are out of date, and it is not feasible (at least for now) for us
to try to reimplement them all.

For now they are marked as dead code. Just delete them to reduce the
amount of code in this library.

Also delete some commented-out code which has hardcoded programs which
are long-obsolete.