dag: remove Witness and Disconnect from abstract DAG
We are no longer using the generic iteration logic to unshare witnesses
or do any other Simplicity-specific analysis. In fact, we have a
parallel PR #142 which uses this `Dag` trait for several things that
aren't even SImplicity programs.

After this commit we may want to consider extending this to handle
thresholds and/or upstreaming to rust-miniscript. (Though it's unclear
that we need threshold support, and unclear that they need DAG (vs tree)
support, so maybe we should just add a separate module there..)