dag: introdue `VerbosePreOrderIter`
The previous commit eliminated recursion by using the new pre-order
iterator, but in a few cases this required a bunch of manual stack
management to keep track of the iterator state.

Encapsulate this stack management in a new `VerbosePreOrderIter` so that
we can write cleaner code.

I think this will ultimately be necessary if we try to do analogous
things to eliminate the recursion in e.g. `Type::finalize`, since there
the recursion is really difficult to emulate.