types: move `occurs_check` from `bound` up to `TypeInner`
The result is theoretically less efficient since we are now doing the
occurs-check on all types, not just Bound ones, but it greatly reduces
the noise caused by the occurs-check, since conceptually this flag
belongs in the Type rather than in a Constraint.

This commit will be superceded by the end of this PR, when we remove the
`occurs_check` boolean entirely, but it is very helpful at this point,
since it will move all the occurs-check logic out of the way of further
refactorings of `Bound` and `Constraint`.