types: move "complete" from Option inside Constraint::Bound to Bound
Now the only difference between Constraint and Bound is that Constraint
has some extra metadata related to the union-bound algorithm. We will
abstract this out in the next commit, and then double-back and work on
eliminating recursion and simplifying these algorithms. (We haven't done
so in these last few commits to try to minimize the diff, both in terms
of LOC and in terms of how the algorithms work. This PR does amount to
another rewrite of the type engine, but it can be understood as a series
of small changes.)