policy: replace PublicKey32 by ToPublicKey and Unsatisfiable
The PublicKey32 trait allows keys to be written to and read from 32-byte
arrays. As written, this can only be implemented for XOnlyPublicKey32
because no other key type in use can round-trip from 32-bytes.

The only place we use the *read* functionality is to load an unspendable
key for use with Taproot. So replace that single use with a dedicated
trait, which maybe we should upstream to rust-miniscript.

For *write* the existing `ToPublicKey` trait covers us, and will let us
use normal key types like xpubs or `DescriptorPublicKey`s.