types: eliminate all recursion from Final and CompleteBound
Note that we still use #derive to get various trait impls for
CompleteBound; this is fine, even though this is a recursive type,
because they immediately pass through to the impls on Final, which are
not recursive.

This commit should also dramatically shrink the debug output of Final,
by only listing the top-level TMR and bitwidth.

It also eliminates the cached `display` variable and just uses dynamic
TMR checks to determine whether we're looking at a power-of-two type.

For some reason this commit brings with it a ~4x speedup for the unit
tests on my machine :).