types: update block comment
The new comment reflects that `Constraint` has been dropped in favor of
`Bound`. This does not actually occur until the next commit, but we add
the comment here because the next commit is large enough as it is, and
it may be helpful for reviewers to take a break here and understand what
the goal of this PR is.