add FromStr and 'static bound to the `Jet` trait
We have implemented `FromStr` on all jet types from the Haskell, code,
but it's a pain to use unless we add the bound to the trait itself.

Also, jets are always just simple enums, which are always 'static. By
making this bound explicit we simplify code that constructs
jet-containing objects.