types: factor out union-bound algorithm
We have now gotten rid of the `Constraint` and `TypeInner` types; we
only have `Bound`, which expresses bounds for incomplete types, which is
held inside the `MutexBound` and `UbElement` containers, which provides
mutable access and union-bound unificaton, respectively, but which don't
contribute any semantic information.

We now have a fairly simple situation where we have the type `Bound`
which is mutally recursive with `Type`.

This also lets us get rid of the code in `unify` that checks whether we
are trying to unify a complete type with anything; this can be done in
one place, in `bind`, which is careful to never modify a complete type.

From here we should be able to eliminate the use of recursive functions
without needing to thread union-bound stuff around or deal with
triply-recursive types where the actual type status is smeared across
all three of them.

This adds a FIXME about a race condition related to the occurs check,
but this race condition already existed, it is not new in this commit.
This is fixed in a later commit.