node: move <J> from Node into node::Marker
We have node::Marker types such as Commit/Construct/Redeem which
conceptually should've had a <J: Jet> parameter, but originally did not
because I also wanted them to be non-constructible, i.e. be an empty
enum, and empty enums can't have parameters.

This led to a massive proliferation of <J> bounds everywhere *else* in
the code, in particular in the form Node<N, J> where "N, J" conceptually
was a single thing.

But then I read

https://stackoverflow.com/questions/75970810/empty-enum-with-generic-parameters

which pointed out that I can totally create non-constructible types with
parameters. I just need to create a struct with an empty enum alongside a
phantomdata.

If we wanted to provided proofs of non-constructibility we could even
add an accessor method to get the `std::convert::Infallible` out, but in
practice I don't think we need such proofs in our code; we just want to
prevent people from accidentally constructing these (and to assure
people that nobody else accidentally constructed one).