remove trait bounds on struct definitions
I left a few in, where they were necessary, and for BitIter and
BitWriter, which seemed uniquely confusing to allow to exist without
the Iterator bound on their children.

Basically this removes a bunch of D: DagLike and J: Jet bounds from
structures, and makes a half-assed effort to remove these bounds from
impls where they are no longer necessary.

This makes our API a little bit more flexible and may reduce boilerplate
for some users (including us). Though the only obviously interesting
consequence is that you can now use `node::Inner` as a generic container
and do manipulations such as `transpose` without needing any bounds on
the contents. In fact, anything you can do on a generic `Inner`, you can
do only with `J: Clone` and `W: Clone` it turns out.

This also serves a documentation/proof purpose, by asserting that none
of the methods on `Inner` actually use the jet-ness or witness-ness of
the J or W fields.

This one might be a little controversial.
See https://stackoverflow.com/a/66369912 for justification.