disconnect: introduce Node::Disconnect type but don't really use it
Introduces the generic Node::Disconnect which the right child of
disconnect will use; set it to Arc<Node<N>> for every node type so
that there is no actual change.

Despite there being no changes, by moving from a concrete type to a
generic one, we need to introduce and use two new crates:

* DisconnectConstructible allows constructing ConstructNodes and
  WitnessNodes even with disconnects
* Disconnectable allows the disconnect nodes to interact with the
  iterators in dag.rs properly.

We also need to fish generic parameters through pretty-much the whole
crate. This is unavoidable, sorry.

This is a super noisy commit but actually has no functionality changes.
The next one, which actually changes the Disconnect types for the
different node types, and which has the real changes in it, should
hopefully be much smaller.