commit: inline decoding logic
The `decode_program` method is now used only in one place, and it is
confusing to have it in the decode module since it's only needed by the
node::commit module.

With this commit, `ConstructNode` becomes the "decoded node" type which
all other nodes need to convert from when they are being decoded.