decode: move decoding unit tests into the appropriate node module
We had a bunch of tests in `decode` that were really testing
`CommitNode::decode` or `RedeemNode::decode` (or both) and had a number
of hacks to make things work when e.g. there was witness data that
CommitNode couldn't handle.

When we change the disconnect parsing for CommitNode vs RedeemNode, we
will find that these hacks are not enough.

Bite the bullet and copy all these tests into the node modules, making
sure that all the unit tests that use `disconnect` are parsed only as
`RedeemNode`s.

We will need to make new unit tests when we introduce the single-child
disconnect nodes for CommitNode, but we'll cross that bridge when we
come to it.