descriptor: Include internal key in serialization
We want to focus on the Simplicity leaves, but Simplicity descriptors
are first and foremost Taproot descriptors with an internal key.
We have the trait UnspendableKey to construct outputs with unspendable
internal key, but for serialization we don't want to be this limited.
In particular, it will be annoying to parse a descriptor with xprivs
if the unspendable internal key requires special treatment because its
private component is unknown.