descriptor: parse_descriptor and to_string_with_secret
This enables us to handle xpriv descriptors.
They don't exist in data, but we can parse to a keymap and an xpub
descriptor. This gives us a guarantee that the keymap corresponds to the
xpub, so it's safe to use during spending.

Copied from rust-bitcoin.

I don't like copy-pasting so many LOC but it seems inevitable.
There is a lot of boilerplate and repetions that could probably be
handled by LISP-style metaprogramming that is not available to us.