simplicity: Update to latest bitcoin and miniscript
bitcoin-hashes no longer provides ToHex and elements is an optional
dependency, so I copied hex.rs from elements, which was itself copied
from the previous version of bitcoin-hashes. Once bitcoin-hashes is
fixed we can easily switch over to that.

Fuzzing is updated.

We have multiple versions of bitcoin_hashes.
slip21 requires bitcoin_hashes < 0.12.0, which resolves to 0.11.
We have to be careful to overwrite 0.12.0 but not 0.11.

The jet benches are updated.

FFI is updated.