policy: Add (Simplicity)Satisfier
The Satisfier from elements-miniscript, but just the methods that we
actually need, using the Simplicity key traits.

Implemented absolute and relative locktime checks because they are
confusing. Made use of the new relative locktime struct from
rust-bitcoin, which we might want to repeat upstream (there is a todo in
the code).

The rest of the satisfier is bare-bones, as miniscript will implement
(Simplicity)Satisfier for (Miniscript)Satisfier, which will bring most
of the convenience. We can extend this later if we want. For now there
are as few changes to rust-simplicity as possible.