CI: Use contrib/test.sh
Mostly copied from rust-bitcoin.

We replace the existing CI with an almost-equivalent.
There is one job per architecture + rust version.
Clippy, rustfmt, benches, simpcli etc. are optional parts of each job.

What is different now is that rustfmt is run on nightly and we added a
rustdoc test. This is what rust-bitcoin does. If it makes sense, we
could switch back to stable.