policy: Add sighash cache
The signature hashes are different for Simplicity than for Taproot.
Among many things, the Simplicity sighash commits to the annexes of all
inputs, while Taproot only commits to the annex of the local input.
There is no clean way to compute the sighash at the moment; the code is
hidden somewhere in the Elements environment in the C files.
We can naively construct a Rust Elements environment in order to extract
the sighash from its FFI bindings. This requires information (control
block) that we arguably don't need, while other information is left
unused (leaf hash). We should write more direct bindings or implement
the sighash in pure Rust in the future. This will take time, so we make
do with what we have.