policy: Iterator over fragments and keys
The order of fragments doesn't seem to matter, so I chose preorder which
is easy to implement.

I would have used the iterators from dag.rs, but Dag doesn't cover nodes
with arbitrary but fixed outdegree, like Policy::Threshold.
It might be worth to extend Dag in the future, although that could make
the generic code in dag.rs more complicated than it already is.