Policy: Add commit and CMR method
We generalize serialize_no_witness to work for CMRs as well.
This makes use of the implementation of Constructible for Cmr.
We introduce the methods commit() and cmr() for users to interact with
Policy. We mark serialize_no_witness() as private because specialized
methods are clearer. In the next commit we will introduce a private type
AssemblyConstructible which will be part of the signature of
serialize_no_witness(). If the method were public, then we would be
forced to make AssemblyConstructible public.