NamedCommitNode: to_witness_node()
This conversion populates
(1) witness nodes based on a hash map from node names to values, and
(2) disconnected branches (holes) based on a forest.

We use the same name space for all nodes, from the main program and from
its disconnected expressions. Each node name must be unique.

We keep empty all witness nodes or holes without a mapping. During the
conversion to WitnessNode, we don't know which branches will be pruned.
The pruning will be done in the conversion to RedeemNode, which will
also throw errors if there is an unpopulated witness node or hole.