Human encoding: Test pruned witness nodes
Witness nodes can remain unpopulated when they are pruned.
to_witness_node keeps witness nodes unpopulated if there is no value in
the witness map.

We currently prune bottom up: Empty witness nodes are marked as to be
pruned. Case nodes with to be pruned branches become assert{l,r}.
THIS IGNORES WHICH CASE BRANCH IS TAKEN BY THE PROGRAM!
A program is executed top-down: Case branches are chosen based on left
or right sum values. Right now, it can easily happen that a case node is
converted into assertr but the left branch is taken. In this case, the
execution fails at a pruned branch.