Cost: Directly pass script witness
This is more flexible because it allows the caller to directly pass a
vector. If the caller wants to reference a txin, they can do so.

We pass &Vec<Vec<u8>> instead of &[Vec<u8>] because
bitcoin::consensus::encode::Encodable is implemented for the former
but not for the latter.