Cost: Separate method to compute budget
Computing the budget in a separate method enables us to compare it with
the required budget. Their difference is the length of the required
padding.

The method panics if the witness stack is larger than u32::MAX bytes.
This assumption existed before this commit, where we casted usize to
u32.

A segwit block is roughly 4 million bytes. u32::MAX is roughly 4
billion bytes, which is much larger. The panic will never fire in
practical scenarios. However, the method would fail in a fuzzer. We
could think about better error handling if it is necessary.