Update jets: Use locally checked-out files
I didn't like how the old script fiddled with my git repos.
What if there are unstaged changes?
What if the target commit is not on the main repo yet?
The script left the repo in an unclean state, with the target commit
still checked out.

Since we have to manage a local git repo anyway, let us also manage the
git revisions. The user has to make sure that the checked-out git
revision of libsimplicity is compatible with the vendored libsimplicity
inside simplicity-sys.