types: use pointer-identity sharing in occurs-check loop
Our existing occurs-check loop iterates through an incomplete type using
the NoSharing sharing tracker. This means that for exponentially-sized
types, we will do exponentially many iterations, even though the types
by-construction have a linear-sized description (in the number of shared
nodes in the program).

What we want to do is use the InternalSharing sharing tracker, *but*
this tracker is too agressive about sharing: it will not repeat nodes at
all, while to do the occurs-check, we want to repeat nodes that occur
inside themselves.

Instead we just implement a pre-order iterator inline (it is very simple
and we don't really need the dag.rs abstractions), implementing exactly
the occurs-check that we need.