types: refactor arrow module
The arrow module has a ton of `for_*` methods which duplicate the
methods in the Constructible traits. Most of these are purely redundant
and can be deleted. A couple provide some code reuse, and are demoted to
be non-`pub`.

Smoe variable names changed in the redundant code, so this will likely
not appear as a code-move in the diff, even though it is one. You can
just skim this commit.