types: introduce inference context object, thread it through the API
This is a super noisy commit, but all it does is introduce the
src/types/context.rs module with a dummy Context type, and make all the
node construction APIs take this and make copies of handles of it.

The next commit will actually _use_ the context in type inference.

This one can therefore be mostly skimmed. The maybe-interesting parts
are the conversion methods, which can be found by searching the diff for
calls to `from_inner`.