types: add &Context to recursive type constructors
The sum and product constructors don't obviously need to be passed a
type context -- after all, the two child types already have a type
context embedded in them.

However, it turns out in practice we have a context available
every single time we call these methods, and it's a little bit
difficult to pull the context out of Types, since Types only
contain a weak pointer to the context object, and in theory
the context might have been dropped, leading to a panic.

Since the user already has a context handle, just make them pass it
in. Then we know it exists.