types: introduce BoundRef type, use in place of Arc<BoundMutex> in union-bound
This introduces the BoundRef type, an opaque reference type which
currently just holds an Arc<BoundMutex>, but which in a later commit
will be tied to the type inference context and refer to data within the
context.

It currently preserves the get() and set() methods on BoundRef, which
come from BoundMutex, but these will need to be replaced in a later
commit, since eventually BoundRef by itself will not contain enough
information to update its data.

This is essentially an API-only change though it does move some data
structures around.